\name{battestdata10}
\docType{data}
\alias{battestdata10}
\title{Tenth Bat Test Data Set }
\description{
This data set gives the barcodes for 89 bats that were held out of battraindata10 to be classified.  The barcode lengths are 659 nucleotides long.
}
\usage{battestdata10}
\format{A data frame containing 89 observations.}
\source{Clare, E. L., B. K. Lim, M. D. Engstrom, J. L. Eger, and P. D. N. Hebert (2006). DNA barcoding of Neotropical bats: species identifcation and discovery within Guyana.}
\keyword{datasets}