\name{battestdata4}
\docType{data}
\alias{battestdata4}
\title{Fourth Bat Test Data Set }
\description{
This data set gives the barcodes for 81 bats that were held out of battraindata4 to be classified.  The barcode lengths are 659 nucleotides long.
}
\usage{battestdata4}
\format{A data frame containing 81 observations.}
\source{Clare, E. L., B. K. Lim, M. D. Engstrom, J. L. Eger, and P. D. N. Hebert (2006). DNA barcoding of Neotropical bats: species identifcation and discovery within Guyana.}
\keyword{datasets}