\name{battestdata7}
\docType{data}
\alias{battestdata7}
\title{Seventh Bat Test Data Set }
\description{
This data set gives the barcodes for 82 bats that were held out of battraindata7 to be classified.  The barcode lengths are 659 nucleotides long.
}
\usage{battestdata7}
\format{A data frame containing 82 observations.}
\source{Clare, E. L., B. K. Lim, M. D. Engstrom, J. L. Eger, and P. D. N. Hebert (2006). DNA barcoding of Neotropical bats: species identifcation and discovery within Guyana.}
\keyword{datasets}