\name{battraindata3}
\docType{data}
\alias{battraindata3}
\title{Third Bat Reference Data Set }
\description{
This data set gives the barcodes for 758 bats from 96 species.  The barcode lengths are 659 nucleotides long.
}
\usage{battraindata3}
\format{A data frame containing 758 observations.}
\source{Clare, E. L., B. K. Lim, M. D. Engstrom, J. L. Eger, and P. D. N. Hebert (2006). DNA barcoding of Neotropical bats: species identifcation and discovery within Guyana.}
\keyword{datasets}