\name{battraindata4}
\docType{data}
\alias{battraindata4}
\title{Fourth Bat Reference Data Set }
\description{
This data set gives the barcodes for 759 bats from 96 species.  The barcode lengths are 659 nucleotides long.
}
\usage{battraindata4}
\format{A data frame containing 759 observations.}
\source{Clare, E. L., B. K. Lim, M. D. Engstrom, J. L. Eger, and P. D. N. Hebert (2006). DNA barcoding of Neotropical bats: species identifcation and discovery within Guyana.}
\keyword{datasets}